+++ src/napari_threedee/_backend/manipulator/_interface.py
