+++ src/napari_threedee/_backend/manipulator/_tests/test_manipulator_visual_data.py
@@ -1,11 +1,11 @@
 import numpy as np
 
-from napari_threedee._infrastructure.manipulator.axis_model import AxisModel
-from napari_threedee._infrastructure.manipulator.central_axis import CentralAxis, CentralAxisSet
-from napari_threedee._infrastructure.manipulator.manipulator_model import ManipulatorModel
-from napari_threedee._infrastructure.manipulator.translator import Translator, TranslatorSet
-from napari_threedee._infrastructure.manipulator.rotator import Rotator, RotatorSet
-from napari_threedee._infrastructure.manipulator.vispy_visual_data import ManipulatorLineData, \
+from napari_threedee._backend.manipulator.axis_model import AxisModel
+from napari_threedee._backend.manipulator.central_axis import CentralAxis, CentralAxisSet
+from napari_threedee._backend.manipulator.manipulator_model import ManipulatorModel
+from napari_threedee._backend.manipulator.translator import Translator, TranslatorSet
+from napari_threedee._backend.manipulator.rotator import Rotator, RotatorSet
+from napari_threedee._backend.manipulator.vispy_visual_data import ManipulatorLineData, \
     ManipulatorHandleData, ManipulatorVisualData
