+++ src/napari_threedee/_backend/manipulator/axis_model.py
