+++ src/napari_threedee/_backend/manipulator/manipulator_model.py
