+++ src/napari_threedee/_backend/manipulator/napari_manipulator.py
@@ -9,12 +9,12 @@ from .axis_model import AxisModel
 from .manipulator_model import ManipulatorModel
 from .vispy_visual_data import ManipulatorVisualData
 from .vispy_manipulator_visual import ManipulatorVisual
-from napari_threedee._infrastructure.manipulator.drag_managers import RotatorDragManager, \
+from napari_threedee._backend.manipulator.drag_managers import RotatorDragManager, \
     TranslatorDragManager
-from napari_threedee._infrastructure.utils.napari_utils import get_vispy_node, \
+from napari_threedee.utils import get_vispy_node, \
     get_mouse_position_in_displayed_layer_data_coordinates, \
     add_mouse_callback_safe, remove_mouse_callback_safe
-from napari_threedee._infrastructure.utils.selection_utils import select_sphere_from_click
+from napari_threedee.utils import select_sphere_from_click
 
 
 class NapariManipulator:
