+++ src/napari_threedee/_backend/manipulator/rotator.py
