+++ src/napari_threedee/_backend/manipulator/utils.py
