+++ src/napari_threedee/_backend/manipulator/vispy_manipulator_visual.py
