+++ src/napari_threedee/_backend/manipulator/vispy_visual_data.py
