+++ src/napari_threedee/_backend/threedee_model.py
