+++ src/napari_threedee/_backend/threedee_widget_model.py
@@ -4,8 +4,8 @@ import napari
 from PyQt5.QtWidgets import QWidget, QPushButton, QVBoxLayout
 from napari.utils.events import Event
 
-from napari_threedee._infrastructure._threedee_base import ThreeDeeModel
-from napari_threedee._infrastructure.utils.napari_utils import generate_populated_layer_selection_widget
+from napari_threedee._backend.threedee_model import ThreeDeeModel
+from napari_threedee.utils import generate_populated_layer_selection_widget
 
 
 class QtThreeDeeWidgetBase(QWidget):
