+++ src/napari_threedee/annotators/spline_annotator.py
@@ -9,9 +9,9 @@ from psygnal import EventedModel
 from pydantic import validator, PrivateAttr
 from scipy.interpolate import splprep, splev
 
-from napari_threedee._infrastructure._threedee_base import ThreeDeeModel
+from napari_threedee._backend.threedee_model import ThreeDeeModel
 from ..mouse_callbacks import add_point_on_plane
-from napari_threedee._infrastructure.utils import add_mouse_callback_safe, remove_mouse_callback_safe
+from napari_threedee.utils import add_mouse_callback_safe, remove_mouse_callback_safe
 
 from typing import Tuple, Union, Optional
