+++ src/napari_threedee/utils/geometry.py
