+++ src/napari_threedee/utils/selection_utils.py
